requirements.txt
fastapi[standard]
uvicorn
requests
python-multipart
jinja2
starlette
python-dotenv 